project/requirements.txt
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.2
yellowbrick>=1.5
plotly>=5.9.0
seaborn>=0.12.0
matplotlib>=3.6.0
pyarrow>=10.0.0
fastparquet>=2023.1.0
streamlit>=1.24.0
kaleido==0.2.1
python-dotenv
requests
